+++ src/main/java/A1B1O3/bodyrecord/auth/domain/repository/TokenRepository.java
@@ -3,5 +3,5 @@ package A1B1O3.bodyrecord.auth.domain.repository;
 import A1B1O3.bodyrecord.auth.domain.Token;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface TokenRepository extends JpaRepository<Token, Long> {
+public interface TokenRepository extends JpaRepository<Token, Integer> {
 }
